+++ CHANGELOG.md
@@ -437,7 +437,7 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - Move from giganto
 
-[Unreleased]: https://github.com/aicers/giganto-client/compare/0.27.0...HEAD
+[Unreleased]: https://github.com/aicers/giganto-client/compare/0.27.0...main
 [0.27.0]: https://github.com/aicers/giganto-client/compare/0.26.0...0.27.0
 [0.26.0]: https://github.com/aicers/giganto-client/compare/0.25.0...0.26.0
 [0.25.0]: https://github.com/aicers/giganto-client/compare/0.24.0...0.25.0